Completed dash board and the page to display all the patients that registered in a given hospital
Patients can register in any hospital of their choice. In this case, a copy of that list is saved in the local database of that hospital and a copy is extended to the cloud. The page that displays all patients registered in a hospital will only display the patients in the local database, ie patients that registered in that hospital. However, any other patients can be retrieved by using the patient's unique ID